Test splitting up project files
