+++ physionet-django/project/models/submission.py
@@ -1,0 +1,270 @@
+
+
+class EditLog(models.Model):
+    """
+    Log for an editor decision. Also saves submission info.
+    """
+    # Quality assurance fields for data and software
+    QUALITY_ASSURANCE_FIELDS = (
+        # 0: Database
+        ('soundly_produced', 'well_described', 'open_format',
+         'data_machine_readable', 'reusable', 'no_phi', 'pn_suitable'),
+        # 1: Software
+        ('soundly_produced', 'well_described', 'open_format', 'no_phi',
+            'reusable', 'pn_suitable'),
+        # 2: Challenge
+        ('soundly_produced', 'well_described', 'open_format',
+         'data_machine_readable', 'reusable', 'no_phi', 'pn_suitable'),
+        # 3: Model
+        ('soundly_produced', 'well_described', 'open_format',
+         'data_machine_readable', 'reusable', 'no_phi', 'pn_suitable'),
+    )
+    # The editor's free input fields
+    EDITOR_FIELDS = ('editor_comments', 'decision', 'auto_doi')
+
+    COMMON_LABELS = {
+        'reusable': 'Does the project include everything needed for reuse by the community?',
+        'pn_suitable': 'Is the content suitable for PhysioNet?',
+        'editor_comments': 'Comments to authors',
+        'no_phi': 'Is the project free of protected health information?',
+        'data_machine_readable': 'Are all files machine-readable?'
+    }
+
+    LABELS = (
+        # 0: Database
+        {'soundly_produced': 'Has the data been produced in a sound manner?',
+         'well_described': 'Is the data adequately described?',
+         'open_format': 'Is the data provided in an open format?',
+         'data_machine_readable': 'Are the data files machine-readable?'},
+        # 1: Software
+        {'soundly_produced': 'Does the software follow best practice in development?',
+         'well_described': 'Is the software adequately described?',
+         'open_format': 'Is the software provided in an open format?'},
+        # 2: Challenge
+        {'soundly_produced': 'Has the challenge been produced in a sound manner?',
+         'well_described': 'Is the challenge adequately described?',
+         'open_format': 'Is all content provided in an open format?'},
+        # 3: Model
+        {'soundly_produced': 'Does the software follow best practice in development?',
+         'well_described': 'Is the software adequately described?',
+         'open_format': 'Is the software provided in an open format?'},
+    )
+
+    HINTS = {
+        'no_phi': [
+            'No dates in WFDB header files (or anonymized dates only)?',
+            'No identifying information of any individual'
+            ' (caregivers as well as patients)?',
+            'No ages of individuals above 89 years?',
+            'No hidden metadata (e.g. EDF headers)?',
+            'No internal timestamps, date-based UUIDs or other identifiers?',
+        ],
+        'open_format': [
+            'No compiled binaries or bytecode?',
+            'No minified or obfuscated source code?',
+        ],
+    }
+
+    content_type = models.ForeignKey(ContentType, on_delete=models.CASCADE)
+    object_id = models.PositiveIntegerField()
+    project = GenericForeignKey('content_type', 'object_id')
+
+    # When the submitting author submits/resubmits
+    submission_datetime = models.DateTimeField(auto_now_add=True)
+    is_resubmission = models.BooleanField(default=False)
+    author_comments = models.CharField(max_length=20000, default='')
+    # Quality assurance fields
+    soundly_produced = models.NullBooleanField(null=True)
+    well_described = models.NullBooleanField(null=True)
+    open_format = models.NullBooleanField(null=True)
+    data_machine_readable = models.NullBooleanField(null=True)
+    reusable = models.NullBooleanField(null=True)
+    no_phi = models.NullBooleanField(null=True)
+    pn_suitable = models.NullBooleanField(null=True)
+    # Editor decision. 0 1 2 for reject/revise/accept
+    decision = models.SmallIntegerField(null=True)
+    decision_datetime = models.DateTimeField(null=True)
+    # Comments for the decision
+    editor_comments = models.CharField(max_length=20000)
+    auto_doi = models.BooleanField(default=True)
+
+    def set_quality_assurance_results(self):
+        """
+        Prepare the string fields for the editor's decisions of the
+        quality assurance fields, to be displayed. Does nothing if the
+        decision has not been made.
+        """
+        if not self.decision_datetime:
+            return
+
+        resource_type = self.project.resource_type
+
+        # See also YES_NO_UNDETERMINED in console/forms.py
+        RESPONSE_LABEL = {True: 'Yes', False: 'No', None: 'Undetermined'}
+
+        # Retrieve their labels and results for our resource type
+        quality_assurance_fields = self.__class__.QUALITY_ASSURANCE_FIELDS[resource_type.id]
+
+        # Create the labels dictionary for this resource type
+        labels = {**self.__class__.COMMON_LABELS, **self.__class__.LABELS[resource_type.id]}
+
+        self.quality_assurance_results = []
+        for f in quality_assurance_fields:
+            qa_str = '{} {}'.format(labels[f], RESPONSE_LABEL[getattr(self, f)])
+            self.quality_assurance_results.append(qa_str)
+
+
+class CopyeditLog(models.Model):
+    """
+    Log for an editor copyedit
+    """
+    content_type = models.ForeignKey(ContentType, on_delete=models.CASCADE)
+    object_id = models.PositiveIntegerField()
+    project = GenericForeignKey('content_type', 'object_id')
+    # Either the time the project was accepted and moved into copyedit
+    # from the edit stage, or the time it was reopened for copyedit from
+    # the author approval stage.
+    start_datetime = models.DateTimeField(auto_now_add=True)
+    # Whether the submission was reopened for copyediting
+    is_reedit = models.BooleanField(default=False)
+    made_changes = models.NullBooleanField(null=True)
+    changelog_summary = models.CharField(default='', max_length=10000, blank=True)
+    complete_datetime = models.DateTimeField(null=True)
+
+
+class LegacyProject(models.Model):
+    """
+    Temporary model for migrating legacy databases
+    """
+    title = models.CharField(max_length=255)
+    slug = models.CharField(max_length=100)
+    abstract = SafeHTMLField(blank=True, default='')
+    full_description = SafeHTMLField()
+    doi = models.CharField(max_length=100, blank=True, default='')
+    version = models.CharField(max_length=20, default='1.0.0')
+
+    resource_type = models.PositiveSmallIntegerField(default=0)
+    publish_date = models.DateField()
+
+    # In case we want a citation
+    citation = models.CharField(blank=True, default='', max_length=1000)
+    citation_url = models.URLField(blank=True, default='')
+
+    contact_name = models.CharField(max_length=120, default='PhysioNet Support')
+    contact_affiliations = models.CharField(max_length=150, default='MIT')
+    contact_email = models.EmailField(max_length=255, default='webmaster@physionet.org')
+
+    # Put the references as part of the full description
+
+    def __str__(self):
+        return ' --- '.join([self.slug, self.title])
+
+    def publish(self, make_file_roots=False):
+        """
+        Turn into a published project
+        """
+        p = PublishedProject.objects.create(title=self.title,
+            doi=self.doi, slug=self.slug,
+            resource_type=ProjectType.objects.get(id=self.resource_type),
+            core_project=CoreProject.objects.create(),
+            abstract=self.abstract,
+            is_legacy=True, full_description=self.full_description,
+            version=self.version,
+            license=License.objects.get(name='Open Data Commons Attribution License v1.0')
+        )
+
+        # Have to set publish_datetime here due to auto_now_add of object
+        dt = datetime.combine(self.publish_date, datetime.min.time())
+        dt = pytz.timezone(timezone.get_default_timezone_name()).localize(dt)
+        p.publish_datetime = dt
+        p.save()
+
+        # Related objects
+        if self.citation:
+            PublishedPublication.objects.create(citation=self.citation,
+                url=self.citation_url, project=p)
+
+        Contact.objects.create(name=self.contact_name,
+            affiliations=self.contact_affiliations, email=self.contact_email,
+            project=p)
+
+        if make_file_roots:
+            os.mkdir(p.project_file_root())
+            os.mkdir(p.file_root())
+
+
+
+class SubmissionInfo(models.Model):
+    """
+    Submission information, inherited by all projects.
+
+    Every project (ActiveProject, PublishedProject, and
+    ArchivedProject) inherits from this class as well as Metadata.
+    The difference is that the fields of this class contain internal
+    information about the publication process; Metadata contains the
+    public information that will be shown on the published project
+    pages.
+
+    In particular, UnpublishedProject.modified_datetime will be
+    updated when any field of Metadata is altered (see
+    UnpublishedProject.save), but not when a field of SubmissionInfo
+    is modified.
+
+    New fields should be added to this class only if they do not
+    affect the content of the project as it will be shown when
+    published.
+    """
+
+    editor = models.ForeignKey('user.User',
+        related_name='editing_%(class)ss', null=True,
+        on_delete=models.SET_NULL, blank=True)
+    # The very first submission
+    submission_datetime = models.DateTimeField(null=True, blank=True)
+    author_comments = models.CharField(max_length=20000, default='', blank=True)
+    editor_assignment_datetime = models.DateTimeField(null=True, blank=True)
+    # The last revision request (if any)
+    revision_request_datetime = models.DateTimeField(null=True, blank=True)
+    # The last resubmission (if any)
+    resubmission_datetime = models.DateTimeField(null=True, blank=True)
+    editor_accept_datetime = models.DateTimeField(null=True, blank=True)
+    # The last copyedit (if any)
+    copyedit_completion_datetime = models.DateTimeField(null=True, blank=True)
+    author_approval_datetime = models.DateTimeField(null=True, blank=True)
+
+    # When the submitting project was created
+    creation_datetime = models.DateTimeField(auto_now_add=True)
+
+    edit_logs = GenericRelation('project.EditLog')
+    copyedit_logs = GenericRelation('project.CopyeditLog')
+
+    # For ordering projects with multiple versions
+    version_order = models.PositiveSmallIntegerField(default=0)
+
+    # Anonymous access
+    anonymous = GenericRelation('project.AnonymousAccess')
+
+    class Meta:
+        abstract = True
+
+    def quota_manager(self):
+        """
+        Return a QuotaManager for this project.
+
+        This can be used to calculate the project's disk usage
+        (represented by the bytes_used and inodes_used properties of
+        the QuotaManager object.)
+        """
+        allowance = self.core_project.storage_allowance
+        published = self.core_project.total_published_size
+        limit = allowance - published
+
+        # DemoQuotaManager needs to know the project's toplevel
+        # directory as well as its creation time (so that files
+        # present in multiple versions can be correctly attributed to
+        # the version where they first appeared.)
+        quota_manager = DemoQuotaManager(
+            project_path=self.file_root(),
+            creation_time=self.creation_datetime)
+        quota_manager.set_limits(bytes_hard=limit, bytes_soft=limit)
+        return quota_manager
+
